+++ fern/__init__.py
@@ -8,5 +8,5 @@
 from fern.logging import Logging
 
 # 初始化一个logging工具, 用户当前模块日志打印
-__logger = Logging('Fern')
+__logger = Logging()
 __logger.add_stream_handler()