+++ fern/logging.py
@@ -12,37 +12,9 @@ from pathlib import Path
 
 class Logging(object):
     """
-    配置日志类
+    配置日志类, 如果是父类 `'.' not in name`, 那么允许继承, 不允许继承
 
-    说明
-    ====
-
-    logging handler继承
-    ----
-    logging支持继承的方式, 例如::
-
-        logger_root = logging.getLogger('root') # or logger_root = logging.getLogger()
-        logger_a = logging.getLogger('a')
-        logger_b = logging.getLogger('a.b')
-
-    在 logger_a 中添加handler_1, 在logger_b中添加 handler_2;
-    执行 logger_b.info('xxx'), 那么它的处理逻辑是:
-
-    - 确认logger_b的level, 如果info < level, 那么下面的步骤都被忽略
-    - 确认logger_b各个handler的level, 如果 info < level, 不做跳过这个handler
-    - 搜索logger_b中的handler并执行
-    - 搜索logger_a中的handler并执行
-    - 搜索logger_root中的handler并执行
-    - 如果前面几个步骤都没有找到handler, 那么会使用 "临时" 创建一个 StreamHandler 并执行
-
-    logging handler 说明
-    ----
-    logging handler 在被创建时, 默认的level值=0, 例如::
-
-        h = logging.StreamHandler()
-        h.level    # 0
-
-    也就是说, 默认情况下, 当前handler不会影响logger的输出
+    由于经常有程序修改 root logger 的配置(如tensorflow), 所有这里会默认屏蔽所有root logger的所有配置
 
     Examples
     ----
@@ -67,10 +39,18 @@ class Logging(object):
         ERROR=logging.ERROR
     )
 
-    def __init__(self, name=__name__, level='info'):
+    def __init__(self, name: str = 'Fern', level: Union[str, int] = 'info', propagate=False):
         logger = logging.getLogger(name)
-        if level.upper() in self.levels:
-            logger.setLevel(self.levels[level.upper()])
+
+        # 所有logger都不从root中继承handler, 但如果是子logger则要从父logger中继承
+        if isinstance(name, str):
+            logger.propagate = ('.' in name)
+        # set level
+        if isinstance(level, int):
+            logger.setLevel(level)
+        else:
+            logger.setLevel(self.levels.get(str(level).upper(), 'WARN'))
+        # add logger
         self.__logger = logger
 
     def __getattr__(self, item: str):